// src/api/generated/routes.ts
// Generated by npm run contract:generate; do not edit.
export const routes = {
  "bulk_edit_objects": {
    "method": "POST",
    "path": "/api/bulk_edit_objects/"
  },
  "correspondents_list": {
    "method": "GET",
    "path": "/api/correspondents/"
  },
  "correspondents_create": {
    "method": "POST",
    "path": "/api/correspondents/"
  },
  "correspondents_partial_update": {
    "method": "PATCH",
    "path": "/api/correspondents/{id}/"
  },
  "correspondents_destroy": {
    "method": "DELETE",
    "path": "/api/correspondents/{id}/"
  },
  "custom_fields_list": {
    "method": "GET",
    "path": "/api/custom_fields/"
  },
  "custom_fields_create": {
    "method": "POST",
    "path": "/api/custom_fields/"
  },
  "custom_fields_partial_update": {
    "method": "PATCH",
    "path": "/api/custom_fields/{id}/"
  },
  "custom_fields_destroy": {
    "method": "DELETE",
    "path": "/api/custom_fields/{id}/"
  },
  "document_types_list": {
    "method": "GET",
    "path": "/api/document_types/"
  },
  "document_types_create": {
    "method": "POST",
    "path": "/api/document_types/"
  },
  "document_types_partial_update": {
    "method": "PATCH",
    "path": "/api/document_types/{id}/"
  },
  "document_types_destroy": {
    "method": "DELETE",
    "path": "/api/document_types/{id}/"
  },
  "documents_list": {
    "method": "GET",
    "path": "/api/documents/"
  },
  "documents_retrieve": {
    "method": "GET",
    "path": "/api/documents/{id}/"
  },
  "documents_partial_update": {
    "method": "PATCH",
    "path": "/api/documents/{id}/"
  },
  "documents_download_retrieve": {
    "method": "GET",
    "path": "/api/documents/{id}/download/"
  },
  "bulk_edit": {
    "method": "POST",
    "path": "/api/documents/bulk_edit/"
  },
  "documents_post_document_create": {
    "method": "POST",
    "path": "/api/documents/post_document/"
  },
  "search_autocomplete_list": {
    "method": "GET",
    "path": "/api/search/autocomplete/"
  },
  "storage_paths_list": {
    "method": "GET",
    "path": "/api/storage_paths/"
  },
  "storage_paths_create": {
    "method": "POST",
    "path": "/api/storage_paths/"
  },
  "storage_paths_partial_update": {
    "method": "PATCH",
    "path": "/api/storage_paths/{id}/"
  },
  "storage_paths_destroy": {
    "method": "DELETE",
    "path": "/api/storage_paths/{id}/"
  },
  "tags_list": {
    "method": "GET",
    "path": "/api/tags/"
  },
  "tags_create": {
    "method": "POST",
    "path": "/api/tags/"
  },
  "tags_partial_update": {
    "method": "PATCH",
    "path": "/api/tags/{id}/"
  },
  "tags_destroy": {
    "method": "DELETE",
    "path": "/api/tags/{id}/"
  },
  "tasks_list": {
    "method": "GET",
    "path": "/api/tasks/"
  }
} as const;